requirements.txt
# ============================================
# SISTEMA DE BIBLIOTECA MUNICIPAL
# Python 3.10+
# ============================================

# ============================================
# CORE FRAMEWORK
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================
# DATABASE & ORM
# ============================================
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
alembic==1.12.1

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ============================================
# CONFIGURATION & ENVIRONMENT
# ============================================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================
# EMAIL & NOTIFICATIONS
# ============================================
aiosmtplib==3.0.1
email-validator==2.1.0

# ============================================
# UTILITIES
# ============================================
python-dateutil==2.8.2

# ============================================
# TESTING (opcional - instalar aparte si necesitas)
# ============================================
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# httpx==0.25.2
# requests==2.31.0

# ============================================
# NOTAS
# ============================================
# Instalar con: pip install -r requirements.txt
# 
# Para testing adicional:
# pip install pytest pytest-asyncio httpx requests
#
# Verificar instalación:
# pip list
#
# Actualizar todas las dependencias:
# pip install --upgrade -r requirements.txt
